Добавлен метод расчёта гравитационного радиуса. Добавлен метод рассчитывающий все гравитационные радиусы в таблице Body

+++ ga-celestial-body/src/main/java/com/klimashin/celestial/body/dto/BodyRequestDto.java
@@ -11,8 +11,8 @@ import lombok.experimental.FieldDefaults;
 public class BodyRequestDto {
 
     String name;
-    float radius;
-    float mass;
-    float gravParameter;
-    float gravRadius;
+    Double radius;
+    Double mass;
+    Double gravParameter;
+    Double gravRadius;
 }
